+++ control-plane/origo_control_plane/definitions.py
@@ -18,6 +18,9 @@ import dagster as dg
 import requests
 from dagster import Definitions
 
+from .assets.daily_okx_spot_trades_to_origo import (
+    insert_daily_okx_spot_trades_to_origo,
+)
 from .assets.daily_trades_to_origo import insert_daily_binance_trades_to_origo
 from .assets.monthly_agg_trades_to_origo import (
     insert_monthly_binance_agg_trades_to_origo,
@@ -49,6 +52,11 @@ insert_monthly_binance_trades_job = define_asset_job(
 insert_daily_binance_trades_job = define_asset_job(
     name='insert_daily_trades_to_origo_job',
     selection=['insert_daily_binance_trades_to_origo'],
+)
+
+insert_daily_okx_spot_trades_job = define_asset_job(
+    name='insert_daily_okx_spot_trades_to_origo_job',
+    selection=['insert_daily_okx_spot_trades_to_origo'],
 )
 
 insert_monthly_binance_agg_trades_job = define_asset_job(
@@ -316,6 +324,7 @@ defs = Definitions(
     assets=[
         insert_monthly_binance_trades_to_origo,
         insert_daily_binance_trades_to_origo,
+        insert_daily_okx_spot_trades_to_origo,
         insert_monthly_binance_agg_trades_to_origo,
         insert_monthly_binance_futures_trades_to_origo,
         insert_monthly_binance_futures_agg_trades_to_origo,
@@ -325,6 +334,7 @@ defs = Definitions(
     jobs=[
         insert_monthly_binance_trades_job,
         insert_daily_binance_trades_job,
+        insert_daily_okx_spot_trades_job,
         insert_monthly_binance_agg_trades_job,
         insert_monthly_binance_futures_trades_job,
         insert_monthly_binance_futures_agg_trades_job,